refactor: `NewTestServerConnection` is a `testutil`
