+++ pkg/client/testutil/serverconnection.go
@@ -1,4 +1,4 @@
-package client_test
+package testutil
 
 import (
 	uclient "github.com/storacha/go-ucanto/client"
@@ -9,14 +9,14 @@ import (
 	uhelpers "github.com/storacha/go-ucanto/testing/helpers"
 )
 
-// newTestServerConnection creates a new Ucanto server and a connection to it. It
+// NewTestServerConnection creates a new Ucanto server and a connection to it. It
 // accepts `server.Option`s to configure the server. This is mainly (if not
 // exclusively) to provide service methods.
 //
 // The server generates its own service principal. It has a `did:web:` DID for
 // realism and readability in errors and failures, but calling code should use
 // `connection.ID()` to get it rather than assume knowledge of the DID it picks.
-func newTestServerConnection(options ...server.Option) uclient.Connection {
+func NewTestServerConnection(options ...server.Option) uclient.Connection {
 	servicePrincipal := uhelpers.Must(signer.Wrap(
 		uhelpers.Must(ed25519.Generate()),
 		uhelpers.Must(did.Parse("did:web:storage.example.com")),
